changed lat and lng from integer to float in the database

+++ db/migrate/20160830160443_create_images.rb
@@ -4,8 +4,8 @@ class CreateImages < ActiveRecord::Migration[5.0]
       t.string :image_url, null: false
       t.text :caption
       t.integer :user_id, null: false, index: true
-      t.integer :lat, null: false
-      t.integer :lng, null: false
+      t.float :lat, null: false
+      t.float :lng, null: false
       t.integer :num_likes, default: 0
       t.timestamps
     end
